+++ Sandbox/Sandbox/LoginPasskeyController.swift
@@ -25,7 +25,7 @@ class LoginPasskeyController: UIViewController {
         
         guard let window = view.window else { fatalError("The view was not in the app's view hierarchy!") }
         AppDelegate.reachfive().login(withRequest: NativeLoginRequest(anchor: window, origin: "LoginPasskeyController.viewDidAppear"), usingModalAuthorizationFor: [.Passkey], display: .IfImmediatelyAvailableCredentials)
-            .onSuccess(callback: goToProfile)
+            .onSuccess(callback: handleLoginFlow)
             .onFailure { error in
                 
                 self.usernameField.isHidden = false
@@ -54,34 +54,37 @@ class LoginPasskeyController: UIViewController {
     
     @IBAction func nonDiscoverableLogin(_ sender: Any) {
         guard let window = view.window else { fatalError("The view was not in the app's view hierarchy!") }
-        let fut: Future<AuthToken, ReachFiveError>
         let request = NativeLoginRequest(anchor: window, origin: "LoginPasskeyController.nonDiscoverableLogin")
+        func onFailure(error: ReachFiveError) -> Void {
+            switch error {
+            case .AuthCanceled:
+                #if targetEnvironment(macCatalyst)
+                    return
+                #else
+                    AppDelegate.reachfive().beginAutoFillAssistedPasskeyLogin(withRequest: NativeLoginRequest(anchor: window, origin: "LoginPasskeyController.nonDiscoverableLogin.AuthCanceled"))
+                        .onSuccess(callback: self.goToProfile)
+                        .onFailure { error in
+                            let alert = AppDelegate.createAlert(title: "Login", message: "Error: \(error.message())")
+                            self.present(alert, animated: true)
+                        }
+                #endif
+            default:
+                let alert = AppDelegate.createAlert(title: "Login", message: "Error: \(error.message())")
+                self.present(alert, animated: true)
+            }
+        }
         switch (usernameField.text) {
         case .none, .some(""):
             // this is optional, but a good way to present a modal with a fallback to QR code for loging using a nearby device
-            fut = AppDelegate.reachfive().login(withRequest: request, usingModalAuthorizationFor: [.Passkey], display: .Always)
+            AppDelegate.reachfive().login(withRequest: request, usingModalAuthorizationFor: [.Passkey], display: .Always)
+                .onSuccess(callback: handleLoginFlow)
+                .onFailure(callback: onFailure)
+            
         case .some(let username):
-            fut = AppDelegate.reachfive().login(withNonDiscoverableUsername: .Unspecified(username), forRequest: request, usingModalAuthorizationFor: [.Passkey], display: .Always)
+            AppDelegate.reachfive().login(withNonDiscoverableUsername: .Unspecified(username), forRequest: request, usingModalAuthorizationFor: [.Passkey], display: .Always)
+                .onSuccess(callback: goToProfile)
+                .onFailure(callback: onFailure)
         }
-        fut.onSuccess(callback: goToProfile)
-            .onFailure { error in
-                switch error {
-                case .AuthCanceled:
-                    #if targetEnvironment(macCatalyst)
-                        return
-                    #else
-                        AppDelegate.reachfive().beginAutoFillAssistedPasskeyLogin(withRequest: NativeLoginRequest(anchor: window, origin: "LoginPasskeyController.nonDiscoverableLogin.AuthCanceled"))
-                            .onSuccess(callback: self.goToProfile)
-                            .onFailure { error in
-                                let alert = AppDelegate.createAlert(title: "Login", message: "Error: \(error.message())")
-                                self.present(alert, animated: true)
-                            }
-                    #endif
-                default:
-                    let alert = AppDelegate.createAlert(title: "Login", message: "Error: \(error.message())")
-                    self.present(alert, animated: true)
-                }
-            }
     }
     
     @IBAction func usernameEditingDidBegin(_ sender: Any) {