+++ Sources/Core/Classes/models/responses/TknMfa.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public class TknMfa: Codable, DictionaryEncodable {
+    public let tkn: String
+    public let mfaRequired: Bool?
+}